requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
googletrans==3.1.0a0
gTTS==2.5.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==0.9.1
httptools==0.6.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.34.1
imageio-ffmpeg==0.4.9
install==1.3.5
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
moviepy==1.0.3
numpy==1.26.4
orjson==3.10.3
packaging==24.0
pillow==10.3.0
proglog==0.1.10
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
pytesseract==0.3.10
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
rfc3986==1.5.0
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
SpeechRecognition==3.10.4
starlette==0.37.2
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.0
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.0
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
